Added more data to the Alert functionality. Created code system to store all alerts via a tree that tracks each exact alert type and success rate

+++ OptionScannerTWS/tWrapper.h
@@ -63,7 +63,7 @@ public:
     double close;
     double high;
     double low;
-    int volume;
+    long volume;
     int barCount = 0;
     double WAP = 0;
     int hasGaps = 0;
